+++ src/eulertools/time.py
@@ -2,6 +2,7 @@ from itertools import product
 
 from eulertools.run import Run
 from eulertools.utils import (
+    ANSIEscape,
     Language,
     Modes,
     Timing,
@@ -30,13 +31,17 @@ class Time:
         self.timings = {language: get_timings(language) for language in languages}
 
     def run(self) -> None:
-        for language, problem in product(self.languages, self.problems):
-            self.time_single_problem(language, problem)
+        for run_index, (language, problem) in enumerate(
+            product(self.languages, self.problems)
+        ):
+            self.time_single_problem(language, problem, run_index)
         if self.run_update:
             for language in self.languages:
                 update_timings(language, self.timings[language])
 
-    def time_single_problem(self, language: Language, problem: str) -> None:
+    def time_single_problem(
+        self, language: Language, problem: str, run_index: int
+    ) -> None:
         self.timings[language].setdefault(problem, {})
         solution = get_solution(language, problem)
         if not solution.exists():
@@ -56,20 +61,45 @@ class Time:
             run_id: Timing.from_nanoseconds(get_average(timings))
             for run_id, timings in raw_timings.items()
         }
-        for key in sorted(new_timings):
+        for key_index, key in enumerate(sorted(new_timings)):
             old_timing = old_timings.get(key)
             raw_timing = raw_timings[key]
             new_timing = new_timings[key]
+            if old_timing is None:
+                overall_difference = None
+            elif old_timing < new_timing:
+                overall_difference = "🔴"
+            elif old_timing > new_timing:
+                overall_difference = "🟢"
+            else:
+                overall_difference = "🔵"
             self.timings[language][problem][key] = new_timing
-            print(f"Old timing: {old_timing}")
-            if self.verbosity > 0:
-                print("New timings:")
-                for i, timing in enumerate(raw_timing):
-                    print(f"* Run {i + 1} took: {Timing.from_nanoseconds(timing)}")
-            prefix = "New" if old_timing is not None else "Initial"
+            title = f"Timing {language.name}/{problem}/{key}..."
+            if run_index or key_index:
+                print()
+            print(ANSIEscape.OKGREEN, title, ANSIEscape.ENDC, sep="")
+            print(ANSIEscape.OKGREEN, "~" * len(title), ANSIEscape.ENDC, sep="")
+            if old_timing:
+                print(f"🟤 Old timing: {old_timing}")
+            prefix = (
+                f"{overall_difference} New" if old_timing is not None else "🔵 Initial"
+            )
             print(f"{prefix} timing: {new_timing}")
+            if self.verbosity > 0:
+                print("    ⏱️  New timings:")
+                for i, timing in enumerate(raw_timing):
+                    run_timing = Timing.from_nanoseconds(timing)
+                    if old_timing is None:
+                        prefix = "🔵"
+                    elif run_timing > old_timing:
+                        prefix = "⬆️ "
+                    elif run_timing < old_timing:
+                        prefix = "⬇️ "
+                    else:
+                        prefix = "↔️ "
+                    print(f"       {prefix} Run {i + 1} took: {run_timing}")
             if self.verbosity > 1 and old_timing is not None:
                 old_nanoseconds = old_timing.nanoseconds
                 new_nanoseconds = new_timing.nanoseconds
                 change = 100 * (old_nanoseconds - new_nanoseconds) / old_nanoseconds
-                print(f"Performance difference: {change:.2f}%")
+                print(f"{overall_difference} Performance difference: {change:.2f}%")